third_party/WebKit/LayoutTests/platform/win/virtual/stable/http/tests/navigation/pushstate-whitelisted-in-blob-denied-expected.txt

This is a testharness.js-based test.
PASS testRunner.addOriginAccessWhitelistEntry is required for this test 
PASS pushState to blob:-URL should fail with SecurityError (even with whitelisted origins) 
Harness: the test ran to completion.
